pyproject.toml
# Copyright 2025 ETH Zurich and University of Bologna.
# Licensed under the Apache License, Version 2.0, see LICENSE for details.
# SPDX-License-Identifier: Apache-2.0
#
# Author: Michael Rogenmoser <michaero@iis.ee.ethz.ch>
[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"

[project]
name = "peakrdl-rawheader"
version = "0.1.3"
requires-python = ">=3.10"
readme = "README.md"
description = "Generate Header files from a SystemRDL register model"
license = {file = "LICENSE"}
dependencies = [
    "mako>=1.2.4",
    "systemrdl-compiler >= 1.29.0, < 2",
    "pytest"
]
authors = [
    {name="Michael Rogenmoser"},
]
keywords = [
    "SystemRDL", "PeakRDL", "CSR", "compiler", "tool", "registers", "generator",
    "C", "header", "software", "SystemVerilog",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]

[project.optional-dependencies]
cli = [
    "peakrdl-cli >= 1.2.3",
]

[project.entry-points."peakrdl.exporters"]
raw-header = "peakrdl_rawheader.__peakrdl__:HeaderGeneratorDescriptor"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.pytest.ini_options]
pythonpath = [
    "peakrdl_rawheader"
]
